native: name every path that leaves a tic unresolved (#492)
New in #493, so it postdated the sweep that widened every other live
test's row struct. Same failure as the others: the server now serves
unresolved as UInt64 and the row's own field still read it as u8.

native_connections_live and native_session_live each build their own
minimal native_state rather than loading the real schema, and both
still declared unresolved UInt8. The poll now reads it as u64, so the
read failed the moment either suite ran against unresolved's own
widened column.

CREATE TABLE IF NOT EXISTS leaves an existing table's columns alone, and
ClickHouse does not fail an insert that narrows a wider value to fit a
stale column: it truncates silently instead. A database loaded before
unresolved widened to UInt64 would read its bits back wrapped modulo
256 rather than erroring, with nothing to say why.

unresolved was one bit: a tic either carried through or it did not, with
no way to tell why short of reading the SQL. unimplemented already names
its own paths this way; unresolved gets the same treatment.

sim::unresolved names thirty bits, one per condition that was already an
independent term in some OR chain feeding the column: a pickup with no
arm, a crossed special line, a damaging sector, a stuck damage call, a
stuck gunshot fold, a weapon cycle that hits an unwritten action or runs
past its own budget, an unfinished player move, a body frame that needs
an action, a used line no manual door handles, a locked door, a stuck
state cycle, a stuck chase, two movers standing too close for either
order to be safe, a missile or skull this does not move, a crossed
special a monster's own allow-list reaches, an unmodeled z-mover, more
than one melee attacker, a stuck attack routine, a kill that counts or
drops an item, a stuck fireball spawn, a melee draw count that could
still undercount, stuck missile thinking, two thinkers sharing a sector,
a stuck plane clip, a floor type that changes a sector's texture, a
plane that reverted, a stuck door thinker, and a crushing plane. A fold
whose own internal cause is not exposed as an independent SQL term at
the point unresolved reads it (the weapon cycle's four cases, the state
cycle's four, the chase's eight, missile thinking's nine) keeps its one
bit rather than being pulled apart, since splitting it means widening a
tuple other callers read too.

mask() folds a list of (bit, condition) pairs into a UInt64 with
bitOr, starting from whatever the tic already carries. Every site that
used to write toUInt8(a OR b OR c) now calls it, so a bit's own comment
is the only place its condition is written out.

unresolved widens to UInt64 in the schema and every reader: the driver's
Refusal, Session's polls, and every live test's own row. refusal.rs's
message now names the bits on both columns, the same "unresolved:
NAME, NAME" shape unimplemented already had.

Live proof, gametic 170 in demo3 (the imp's fireball, thrown the tic
before, is still going and neither the general movers' own stage nor
its Z-axis counterpart runs one for it, and its own momentum stands it
close enough to another momentum mover to crowd that too): the row
reads TX_CROWDED, TX_UNRUN, TZ_UNRUN, and the driver's message names
all three.

+++ native/schema.sql
@@ -831,11 +831,12 @@ CREATE TABLE IF NOT EXISTS {{DB}}.native_state
     -- Interactive input carry
     turnheld  Int32,
 
-    -- A tic the simulation could not produce in full. `unresolved` is 1
-    -- when any field is a placeholder; `unimplemented` names which paths
-    -- the tic reached but did not run, one bit each. Both are zero on a
-    -- tic that ran completely.
-    unresolved     UInt8,
+    -- A tic the simulation could not produce in full. `unresolved` names
+    -- which paths this tic reached but could not run, one bit each;
+    -- `unimplemented` names which paths the level itself carries that
+    -- native mode does not model at all, also one bit each. Both are zero
+    -- on a tic that ran completely.
+    unresolved     UInt64,
     unimplemented  UInt64,
 
     -- What the tic drew from the random tables, in draw order: `dbg_ran`
